db models/helpers refactor
