+++ server/src/database/index.ts
@@ -1,0 +1,3 @@
+export { commentsCollection } from "./collections";
+
+export type { DbComment } from "./models";
