+++ server/src/database/models/index.ts
@@ -1,0 +1,1 @@
+export type { DbComment } from "./DbComment";
